esri_json

Module geo

Enables convertion between Esri geometry types and geometries from [geo_types]

Coord

[geo_types::Coord] can be converted into any of Coord ( CoordXy, CoordXyz, CoordXym, CoordXyzm )

Missing Z coordinates are filled with 0.0_f64 and missing M coordinates are filled with None

Coord can be converted into [geo_types::Coord]

M and Z coordinates are ignored

These changes apply to the coordinates when converting all other geometries

Points

Point types are easily convertible
[geo_types::Point] and Point can be converted to each other

[geo_types::MultiPoint] and MultiPoint can be converted to each other

Lines

[geo_types::Line], [geo_types::LineString] and [geo_types::MultiLineString] can be converted to LineString

LineString can only be converted to [geo_types::MultiLineString]

Polygons

[geo_types::Polygon], [geo_types::MultiPolygon], [geo_types::Rect] and [geo_types::Triangle] can be converted to Polygon

Polygon can only be converted to [geo_types::MultiPolygon]

Geometry

[geo_types::Geometry] can mostly be converted to Geometry, but [geo_types::GeometryCollection] cannot be converted

Geometry can only be converted to [geo_types::Geometry]

Spatial Reference

The SpatialReference type is ignored when converting and will need to be set separately