esri_json

Module geometry

Geometry Primitives representing ESRI geometries

Structs

CoordXy

Base Coordinate type with X and Y coordinates

CoordXym

Base Coordinate type with X and Y coordinates and an optional Measure value

CoordXyz

Base Coordinate type with X, Y and Z coordinates

CoordXyzm

Base Coordinate type with X, Y and Z coordinates and an optional Measure value

LineString

An ordered Collection of Coord, used as part of Polyline and Polygon

MultiPoint

Representation of a MultiPoint

Point

Representation of a Point

Polygon

Representation of a Polygon

Polyline

Representation of a Polyline

SpatialReference

Representation of a SpatialReference

Enums

Geometry

Union type of all Geometry types

Traits

Coord

Coord is the primitive type for all coordinate types It has no spatial reference Optional Z and M values