esri_json::geometry

Trait Coord

pub trait Coord: Clone {
    // Required methods
    fn has_z() -> bool;
    fn has_m() -> bool;
    fn x(&self) -> f64;
    fn y(&self) -> f64;
    fn from_coord<C: Coord>(c: C) -> Self;
    fn from_coord_fields(x: f64, y: f64, z: Option<f64>, m: Option<f64>) -> Self;

    // Provided methods
    fn z(&self) -> Option<f64> { ... }
    fn m(&self) -> Option<f64> { ... }
}

Coord is the primitive type for all coordinate types It has no spatial reference Optional Z and M values

Required Methods

fn has_z() -> bool

fn has_m() -> bool

fn x(&self) -> f64

fn y(&self) -> f64

fn from_coord<C: Coord>(c: C) -> Self

fn from_coord_fields(x: f64, y: f64, z: Option<f64>, m: Option<f64>) -> Self

Provided Methods

fn z(&self) -> Option<f64>

fn m(&self) -> Option<f64>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.