esri_json::webmap

Module esri_sls

Representation of a Simple Line Symbol

EsriSLS and re-exported types used to represent a Simple Line Symbol

Structs

EsriSLS

Representation of a Simple Line Symbol

Marker

https://developers.arcgis.com/web-map-specification/objects/marker/

Enums

Color

A color is either RGB or RGBA

Placement

Style

Representation of the line style of a Simple Line Symbol