esri_json::webscene::layers

Module fill_symbol_3d_layer

Representation of a FillSymbol3DLayer

FillSymbol3DLayer is used to render the surfaces of flat 2D Polygon geometries and 3D volumetric meshes in a SceneView.

Includes re-exports of relevant types

Structs

Edges

Representation of Sketch Edge and Solid Edge

FillSymbol3DLayer

Representation of a FillSymbol3DLayer

Material

Representation of a MaterialColorMixMode

Outline

Representation of a Outline

Enums

Color

A color is either RGB or RGBA

ColorMixMode

Controls how symbolLayer or visualVariable color is applied onto the underlying geometry color/texture.

LineCap

Shape of the tips at the start and end of each line geometry. This also applies to the tips of each pattern segment along the line

LinePattern

PolygonPattern

Representation of a polygonPattern