esri_json::webscene::layers

Module line_symbol_3d_layer

Representation of a LineSymbol3DLayer

LineSymbol3DLayer renders Polyline geometries using a flat 2D line with a LineSymbol3D in a 3D SceneView.

Includes re-exports of relevant types

Structs

LineMarker

Represents markers placed at the start and end of each line geometry, or both

LineSymbol3DLayer

Representation of a LineSymbol3DLayer

Material

The material used to shade the geometry

Enums

Color

A color is either RGB or RGBA

LineCap

Shape of the tips at the start and end of each line geometry. This also applies to the tips of each pattern segment along the line

LineJoin

Shape of the intersection of two line segments

LinePattern

LineSymbolMarkerStyle

Style of the marker

Placement

Indicates where the marker is placed.

PxOrPt

Enum which represents a field which can be Pixels, Points or a number (treated as points)