esri_json::webscene::layers

Module path_symbol_3d_layer

Representation of a PathSymbol3DLayer

PathSymbol3DLayer renders polyline geometries by extruding a 2D profile along the line, resulting in visualizations like tubes, walls, etc.

Includes re-exports of relevant types

Structs

Material

The material used to shade the geometry

PathSymbol3DLayer

Representation of a PathSymbol3DLayer

Enums

Anchor

The position of the extrusion profile with respect to the polyline geometry.

Cap

Shape of the tips at the start and end of each path geometry

Color

A color is either RGB or RGBA

LineJoin

Shape of the intersection of two line segments

Profile

The shape which is extruded along the line

ProfileRotation

Specifies the axes about which the profile may be rotated at the joins. Constraining the rotation axes leads to a fixed orientation of the profile for the specified directions